[Bronze II] Title: 숫자, Time: 848 ms, Memory: 30052 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -3,8 +3,8 @@ import java.util.*;
 public class Main{
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
-        long a = sc.nextInt();
-        long b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         if(a>b){
             long temp = a;
             a = b;
